Moving secrets to env variables
