+++ .github/workflows/deploy.yml
@@ -2,7 +2,7 @@ name: Deploy Vite app to GitHub Pages
 
 on:
   push:
-    branches: ["master", "main"]   # or "master", depending on your default branch
+    branches: ["master"]   # use your default branch
 
 permissions:
   contents: read
@@ -12,6 +12,8 @@ permissions:
 jobs:
   deploy:
     runs-on: ubuntu-latest
+    env:
+      VITE_SPOTIFY_CLIENT_ID: ${{ secrets.SPOTIFY_CLIENT_ID }}
     steps:
       - name: Checkout repo
         uses: actions/checkout@v4
@@ -34,3 +36,5 @@ jobs:
 
       - name: Deploy to GitHub Pages
         uses: actions/deploy-pages@v4
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
