+++ index.html
@@ -10,7 +10,7 @@
 <aside class="leftSidebar"></aside>
 
 <a id="profile" href="https://accounts.spotify.com/login">
-        <img id="imgUrl" src="default-avatar.png" alt="Spotify Profile" />
+        <img id="imgUrl" src="public/default-avatar.png" alt="Spotify Profile" />
 </a>
 
 <section class="currentlyPlaying">
