+++ src/Spotify/authCodeWithPkce.ts
@@ -1,9 +1,7 @@
-// ----------------- Redirect URI Handling -----------------
-// Automatically pick redirect URI based on environment
 const redirectUri =
     import.meta.env.MODE === "development"
-        ? "http://127.0.0.1:5173/spotifyQueue/"  // dev localhost
-        : "https://matyaskrizek.github.io/spotifyQueue/"; // production GitHub Pages
+        ? "http://127.0.0.1:5173/spotifyQueue/spotify"  // dev localhost
+        : "https://matyaskrizek.github.io/spotifyQueue/spotify"; // production GitHub Pages
 
 export async function redirectToAuthCodeFlow(clientId: string): Promise<void> {
     const verifier = generateCodeVerifier(128);
